Généricisation du composant AddressSearch
